refactor(state): rename instruction/ folder to field/
The instruction/ contract (def/get/set/ref return a token expanded at
init) didn't fit hot(), which is an eager reactive-value wrapper. Group
these by call-site shape instead - all are class-field declarations - so
hot() belongs alongside them. 'instruction' stays the public concept noun
in prose and docs; only the folder/file paths move to field/.

Updates skill doc links (SKILL.md, types.md, state.md, context.md) to the
new path.
