+++ packages/state/src/index.ts
@@ -1,8 +1,8 @@
-export { def } from './instruction/def';
-export { get } from './instruction/get';
-export { hot } from './instruction/hot';
-export { set } from './instruction/set';
-export { ref } from './instruction/ref';
+export { def } from './field/def';
+export { get } from './field/get';
+export { hot } from './field/hot';
+export { set } from './field/set';
+export { ref } from './field/ref';
 
 export { State, State as default, unbind } from './state';
 export { watch, listener, event, Observable, observer, touch } from './observable';
